services/ai/pyproject.toml
[project]
name = "cobble-ai"
version = "0.1.0"
description = "Cobble AI FastAPI service"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "python-dotenv==1.0.1",
  "httpx==0.27.2",
  "reportlab==4.2.2",
  "PyNiteFEA==1.3.0",
  "numpy==1.26.4"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "pytest-asyncio==0.23.8",
  "httpx==0.27.2"
]

[tool.pytest.ini_options]
addopts = "-q"

